refactor: Move utils from `src/shared/utils` to their places of usage; Rename `src/checkers` to `src/run-checkers`
