+++ src/run-checkers/checkers/forbidden-checker/index.ts
@@ -1,0 +1,1 @@
+export { forbiddenChecker } from "./forbidden-checker.js";
