+++ src/run-checkers/checkers/index.ts
@@ -1,0 +1,3 @@
+export { forbiddenChecker } from "./forbidden-checker/index.js";
+export { recentChecker } from "./recent-checker/index.js";
+export { semverChecker } from "./semver-checker/index.js";
