+++ src/run-checkers/checkers/recent-checker/index.ts
@@ -1,0 +1,1 @@
+export { recentChecker } from "./recent-checker.js";
