+++ src/run-checkers/checkers/recent-checker/utils/index.ts
@@ -1,0 +1,3 @@
+export { getPackageVersions } from "./get-package-versions.js";
+export { getRecentVersions } from "./get-recent-versions.js";
+export { parseRecentVersions } from "./parse-recent-versions.js";
