+++ src/run-checkers/checkers/semver-checker/index.ts
@@ -1,0 +1,1 @@
+export { semverChecker } from "./semver-checker.js";
