+++ src/run-checkers/checkers/semver-checker/semver-checker.ts
@@ -1,16 +1,16 @@
 import { satisfies } from "semver";
 
-import type { DependenciesInstalled } from "src/shared/utils/index.js";
 import { isArrayOfArrays } from "src/shared/utils/index.js";
 
-import type { CliOptions } from "../command.js";
+import type { CliOptions } from "../../../command.js";
 import type {
   DependencyName,
   SemverRule,
   SemverRuleset,
-} from "../config/index.js";
-import { Severity } from "../config/index.js";
-import { SemverRuleViolation, stats } from "../stats/index.js";
+} from "../../../config/index.js";
+import { Severity } from "../../../config/index.js";
+import { SemverRuleViolation, stats } from "../../../stats/index.js";
+import type { DependenciesInstalled } from "../../utils/index.js";
 
 const checkSemverRule = (
   dependenciesInstalled: DependenciesInstalled,
