+++ src/run-checkers/index.ts
@@ -1,0 +1,1 @@
+export { runCheckers } from "./run-checkers.js";
