+++ src/run-checkers/utils/index.ts
@@ -1,0 +1,3 @@
+export type { DependenciesInstalled } from "./get-dependencies-installed.js";
+export { getDependeniesInstalled } from "./get-dependencies-installed.js";
+export { getDependencyTree } from "./get-dependency-tree.js";
