--- src/shared/utils/npm/index.ts
@@ -1,2 +1,0 @@
-export { getDependencyTree } from "./get-dependency-tree.js";
-export { getPackageVersions } from "./get-package-versions.js";
